Fix copy-electron-to-dist script to check if there is dist dir

+++ copy-electron-to-dist.js
@@ -3,7 +3,20 @@ const fs = require('fs')
 const FROM = `${__dirname}/src/js/electron.js`
 const TO = `${__dirname}/dist/electron.js`
 
-fs.copyFile(FROM, TO, (err) => {
-  if (err) throw err
-  console.log('electron.js was copied to /dist/electron.js')
-})
+const distExists = fs.existsSync(`${__dirname}/dist}`)
+
+if (distExists) {
+  fs.copyFile(FROM, TO, (err) => {
+    if (err) throw err
+    console.log('electron.js was copied to /dist/electron.js')
+  })
+} else {
+  fs.mkdir(`${__dirname}/dist`, {}, err => {
+    if (err) throw err
+    fs.copyFile(FROM, TO, (err) => {
+      if (err) throw err
+      console.log('electron.js was copied to /dist/electron.js')
+    })
+  })
+}
+
